refactor(core): excise the v1 rules-parser assembly + arbitration bridge
Delete the parser-exclusive machinery: core/parser (AddressParser + solution-to-proposals),
core/solvers, core/filters, the #478 arbitration bridge (decoder/arbitrate-tree,
policy/input-shape-router, policy/from-config, the `arbitrate` flag + `ruleProposer` stage in
core/pipeline). The generic proposal collect/filter helpers move to core/policy/proposal-collection.ts
(the neural --policy CLI path still uses them). core/package.json drops the ./parser/./solvers/./filters
export subpaths (both maps); the vitest aliases drop the stale globs and add the missing single-file
core subpath entries (fs/objects/coarse-placer/kysley) so `vitest --root core` resolves from a clean
checkout.

DEVIATION FROM BRIEF (reported): core/tokenization/{context,Graph,permutate} and core/solver are
RETAINED, not deleted. `Span extends Graph`, and the kept @mailwoman/classifiers package →
core/classification (BaseClassifier) → TokenContext (context.ts) + Span → Graph, and context → solver.
The brief's "classifiers STAYS (plan 5)" and "delete context/Graph" are mutually exclusive; keeping the
package (repeated operator ruling) requires keeping its substrate. Full removal lands in plan 5 with the
classifiers workspace.
