--- core/filters/index.ts
@@ -1,10 +1,0 @@
-/**
- * @copyright Sister Software
- * @license AGPL-3.0
- * @author Teffen Ellis, et al.
- */
-
-export * from "./InvalidSolutionFilter.ts"
-export * from "./RelationshipFilter.ts"
-export * from "./SubsetFilter.ts"
-export * from "./TokenDistanceFilter.ts"
