+++ core/index.ts
@@ -8,7 +8,6 @@ export * from "./classification/index.ts"
 export * from "./collections.ts"
 export * from "./decoder/index.ts"
 export * from "./identifiers.ts"
-export * from "./parser/index.ts"
 export * from "./pipeline/index.ts"
 export * from "./resources/index.ts"
 export * from "./solver/index.ts"
