--- core/parser/index.ts
@@ -1,9 +1,0 @@
-/**
- * @copyright Sister Software
- * @license AGPL-3.0
- * @author Teffen Ellis, et al.
- */
-
-export * from "./AddressParser.ts"
-export * from "./proposal-pipeline.ts"
-export * from "./solution-to-proposals.ts"
