+++ core/pipeline/runtime-pipeline.test.ts
@@ -11,8 +11,6 @@ import { describe, expect, it, vi } from "vitest"
 
 import type { AddressNode, AddressTree } from "../decoder/types.ts"
 import type { Resolver } from "../resolver/types.ts"
-import type { Span } from "../tokenization/index.ts"
-import type { ClassificationProposal, ComponentTag } from "../types/index.ts"
 import { HARD_PLACE_COUNTRY_SAFELIST, hardCountryFor, isBareLocalityTree, runPipeline } from "./runtime-pipeline.ts"
 import { WORD_CONSISTENCY_SHIP_DEFAULT } from "./types.ts"
 import type {
@@ -676,128 +674,5 @@ describe("runPipeline — coarse-placer soft prior (#244)", () => {
 		})
 		expect(result.path).toBe("fast-path")
 		expect(seen[0]).toMatchObject({ anchorPosterior: { US: 0.96 }, anchorWeight: 1.0 })
-	})
-})
-
-describe("runPipeline — arbitration (#478 inc 3)", () => {
-	function ruleProp(component: ComponentTag, body: string, start: number, confidence: number): ClassificationProposal {
-		return {
-			span: { start, end: start + body.length, body } as unknown as Span,
-			component,
-			confidence,
-			source: "rule",
-			source_id: "v0-test",
-			penalty: 0,
-		}
-	}
-
-	const cleanKind: QueryKindResult = { kind: "structured_address", confidence: 0.9, alternatives: [] }
-	const alphaShape = (): QueryShapeLite => ({ knownFormats: [], characterClass: "alpha" })
-
-	it("arbitrate unset → neural tree unchanged, ruleProposer never called (byte-stable)", async () => {
-		const neuralRoots: AddressNode[] = [
-			{ tag: "street", value: "Main Street", start: 0, end: 11, confidence: 0.8, children: [] },
-		]
-		const ruleProposer = vi.fn(async () => [ruleProp("street", "Main", 0, 0.9)])
-		const result = await runPipeline("Main Street", {
-			computeQueryShape: alphaShape,
-			classifyKind: async () => cleanKind,
-			classifier: fakeClassifier(fakeTree("Main Street", neuralRoots)),
-			ruleProposer,
-		})
-		expect(ruleProposer).not.toHaveBeenCalled()
-		expect(result.tree.roots).toEqual(neuralRoots)
-		expect(result.timing["arbitrate"]).toBeUndefined()
-	})
-
-	it("rule_preferred PRESERVES neural's street decomposition (the precondition fix)", async () => {
-		// neural: street[4,12]"Seminary" + street_suffix[13,15]"Dr"; rule: combined street[4,15]. fix-v1
-		// never restructures, so both neural nodes survive and the coarse rule street is ignored — the
-		// suffix is no longer able to evict the street (the leg-2 bug).
-		const ruleProposer = vi.fn(async () => [
-			ruleProp("house_number", "109", 0, 1),
-			ruleProp("street", "Seminary Dr", 4, 0.82),
-		])
-		const result = await runPipeline(
-			"109 Seminary Dr",
-			{
-				computeQueryShape: alphaShape,
-				classifyKind: async () => cleanKind,
-				classifier: fakeClassifier(
-					fakeTree("109 Seminary Dr", [
-						{ tag: "house_number", value: "109", start: 0, end: 3, confidence: 0.93, children: [] },
-						{ tag: "street", value: "Seminary", start: 4, end: 12, confidence: 0.94, children: [] },
-						{ tag: "street_suffix", value: "Dr", start: 13, end: 15, confidence: 0.94, children: [] },
-					])
-				),
-				ruleProposer,
-			},
-			{ arbitrate: true }
-		)
-		const tags = result.tree.roots.map((r) => r.tag)
-		expect(tags).toContain("street")
-		expect(tags).toContain("street_suffix") // NOT evicted
-		expect(result.tree.roots.find((r) => r.tag === "street")?.value).toBe("Seminary") // neural decomposition kept
-		expect(result.timing["arbitrate"]).toBeGreaterThanOrEqual(0)
-	})
-
-	it("rule_preferred RELABELS a same-span tag disagreement toward rule", async () => {
-		// neural labels [0,3] locality; rule says region — same span, different tag → rule wins.
-		const ruleProposer = vi.fn(async () => [ruleProp("region", "NYC", 0, 1)])
-		const result = await runPipeline(
-			"NYC",
-			{
-				computeQueryShape: alphaShape,
-				classifyKind: async () => cleanKind,
-				classifier: fakeClassifier(
-					fakeTree("NYC", [{ tag: "locality", value: "NYC", start: 0, end: 3, confidence: 0.6, children: [] }])
-				),
-				ruleProposer,
-			},
-			{ arbitrate: true }
-		)
-		expect(result.tree.roots).toHaveLength(1)
-		expect(result.tree.roots[0]?.tag).toBe("region")
-		expect(result.tree.roots[0]?.source).toBe("rule")
-	})
-
-	it("rule_preferred ADDS a rule-only missing tag on a non-overlapping span", async () => {
-		const ruleProposer = vi.fn(async () => [ruleProp("country", "USA", 10, 1)])
-		const result = await runPipeline(
-			"Main St   USA",
-			{
-				computeQueryShape: alphaShape,
-				classifyKind: async () => cleanKind,
-				classifier: fakeClassifier(
-					fakeTree("Main St   USA", [
-						{ tag: "street", value: "Main St", start: 0, end: 7, confidence: 0.9, children: [] },
-					])
-				),
-				ruleProposer,
-			},
-			{ arbitrate: true }
-		)
-		const country = result.tree.roots.find((r) => r.tag === "country")
-		expect(country?.value).toBe("USA")
-		expect(country?.source).toBe("rule")
-	})
-
-	it("neural_preferred (OOD script) passes the neural tree through unchanged", async () => {
-		const ruleProposer = vi.fn(async () => [ruleProp("street", "abc", 0, 0.95)])
-		const neuralRoots: AddressNode[] = [
-			{ tag: "street", value: "abc", start: 0, end: 3, confidence: 0.5, children: [] },
-		]
-		const result = await runPipeline(
-			"abc",
-			{
-				computeQueryShape: () => ({ knownFormats: [], characterClass: "cjk" }),
-				classifyKind: async () => cleanKind,
-				classifier: fakeClassifier(fakeTree("abc", neuralRoots)),
-				ruleProposer,
-			},
-			{ arbitrate: true }
-		)
-		expect(ruleProposer).not.toHaveBeenCalled()
-		expect(result.tree.roots).toEqual(neuralRoots)
 	})
 })
