--- core/solvers/index.ts
@@ -1,14 +1,0 @@
-/**
- * @copyright Sister Software
- * @license AGPL-3.0
- * @author Teffen Ellis, et al.
- */
-
-export * from "./ExclusiveCartesianSolver.ts"
-export * from "./HouseNumberPositionPenalty.ts"
-export * from "./LeadingAreaDeclassifier.ts"
-export * from "./MultiStreetSolver.ts"
-export * from "./OrphanedLevelTypeDeclassifier.ts"
-export * from "./OrphanedUnitTypeDeclassifier.ts"
-export * from "./PostcodePositionPenalty.ts"
-export * from "./VenueCaptureSolver.ts"
