+++ core/vitest.config.ts
@@ -7,8 +7,10 @@
  *   package.json `exports` field points at `./out/.../index.js` — those files don't exist until tsc
  *   has run. With these aliases vitest can run from a clean checkout.
  *
- *   `solvers/*` and `filters/*` use a glob-style export, so the alias mirrors that with a regex that
- *   preserves the trailing path segment.
+ *   A few subpaths are single files, not directories with an `index.ts` — `fs` is the node build
+ *   (`fs/node.ts`), `objects` is a bare file (`objects.ts`), `coarse-placer`/`kysley/*` are single
+ *   files. Those get explicit entries BEFORE the generic `<subpath> → <subpath>/index.ts` rule,
+ *   which would otherwise resolve them to a non-existent `index.ts` (mirrors the root config).
  */
 
 import { resolve } from "node:path"
@@ -23,10 +25,13 @@ const here = import.meta.dirname
 export default defineConfig({
 	resolve: {
 		alias: [
-			// Order matters — more specific entries first.
-			{ find: /^@mailwoman\/core\/solvers\/(.+)$/, replacement: resolve(here, "solvers/$1") },
-			{ find: /^@mailwoman\/core\/filters\/(.+)$/, replacement: resolve(here, "filters/$1") },
-			// Every @mailwoman/core/<subpath> resolves to the index.ts of the matching subdir.
+			// Order matters — more specific entries first. Single-file subpaths (no directory index)
+			// must beat the generic `<subpath>/index.ts` rule below.
+			{ find: /^@mailwoman\/core\/kysley\/(.+)$/, replacement: resolve(here, "kysley/$1.ts") },
+			{ find: /^@mailwoman\/core\/coarse-placer$/, replacement: resolve(here, "coarse-placer/coarse-placer.ts") },
+			{ find: /^@mailwoman\/core\/objects$/, replacement: resolve(here, "objects.ts") },
+			{ find: /^@mailwoman\/core\/fs$/, replacement: resolve(here, "fs/node.ts") },
+			// Every other @mailwoman/core/<subpath> resolves to the index.ts of the matching subdir.
 			{ find: /^@mailwoman\/core\/(.+)$/, replacement: resolve(here, "$1/index.ts") },
 			{ find: /^@mailwoman\/core$/, replacement: resolve(here, "index.ts") },
 			// Sibling workspaces.
